. blkalloc.h, dir.cpp, file.cpp, regfs.cpp, regfuse.cpp: fixed the bug file or dir could be removed when it was opened by someone else. And now it can be removed from the directory when it is open and get actually removed after all the users close the file, as the other filesystem does.

+++ monitor/registry/regfuse.cpp
@@ -190,6 +190,12 @@ static int regfs_unlink(const char *path)
 {
     stdstr strPath = path;
     CRegistryFs* pfs = g_pRegfs;
+    if( strcmp( path, "/" ) == 0 )
+    {
+        DebugPrint( -EACCES, "Error, cannot "
+            "remove root directory." );
+        return -EACCES;
+    }
     return  pfs->RemoveFile( strPath );
 }
 
@@ -197,6 +203,12 @@ static int regfs_rmdir(const char *path)
 {
     stdstr strPath = path;
     CRegistryFs* pfs = g_pRegfs;
+    if( strcmp( path, "/" ) == 0 )
+    {
+        DebugPrint( -EACCES, "Error, cannot "
+            "remove root directory." );
+        return -EACCES;
+    }
     return  pfs->RemoveDir( strPath );
 }
